Fix unit tests to give ok/not ok
git-svn-id: svn+ssh://svn.code.sf.net/p/eidors3d/code/trunk@3840 82006732-fb7e-4a22-9f28-86a212cce563

+++ eidors/tools/unit_test_cmp.m
@@ -18,7 +18,7 @@ tolstr='';
       if isnan(a) == isnan(b);
           a(isnan(a))=0; b(isnan(b))=0;
        end;
-      if all(abs(a - b) <= tol);
+      if all(abs(double(a) - double(b)) <= tol);
          ok='ok';
       end;
